Projekt_2/requirements.txt
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
minio==7.2.7
boto3==1.34.131
pymongo==4.7.2
streamlit==1.36.0
plotly==5.22.0
pyarrow==16.1.0
python-dotenv==1.0.1
selenium==4.11.2
webdriver-manager==3.8.6
matplotlib==3.8.0
